requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
openai==1.40.6
mistralai==0.4.2
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.9.4
typer==0.12.3
feedparser==6.0.11